+++ src/callable/operators.rs
@@ -380,6 +380,12 @@ impl Callable for PostfixVecIndex {
         let (what, index) = stack.eval_binary(args.unnamed_binary_args())?;
         what.try_get(index)
     }
+    fn call_mutable(&self, args: ExprList, stack: &mut CallStack) -> EvalResult {
+        let x = args.unnamed_binary_args();
+        let what = stack.eval_mut(x.0)?;
+        let index = stack.eval(x.1)?;
+        what.try_get(index)
+    }
 }
 
 #[derive(Debug, Clone, PartialEq)]